requirements.txt
Flask==3.0.0
pytest==8.0.0
requests
flask-socketio
Flask-Mail
python-dotenv
psycopg2-binary
